Refactor: Обновляет тип ID для deleteNote
- Изменяет тип параметра `id` метода `deleteNote` с `int` на `long`.
- Позволяет обрабатывать более широкий диапазон идентификаторов.
- Обеспечивает согласованность с типами ID в базе данных.

+++ src/main/java/com/monolatte/kontur/model/SQLManager/ISQLManager.java
@@ -6,7 +6,7 @@ public interface ISQLManager<T> extends ISQLManagerBase {
 
     void addNote(T note);
 
-    void deleteNote(int id);
+    void deleteNote(long id);
 
     void updateNote(T note);
